Czech translation + fixes of resourses and link generation (#432)
### Translation
This is draft PR for Czech translation of [publiccode.eu](https://publiccode.eu) website. Translation is done and now waits for comments from proofreaders. After proofreading is finished I will remove *DO NOT MERGE* sign.

Status of proofreaders:
- **David** - Done
- **Honza** - Done
- **Klára** - Done
- **Tibor** - Done
- **Kubík** - Done
- **Kačka** - Done
- **Medvěd** - Done

### Links and resources
In this PR also contains fixes of links and resources. This is main reason why I did this draft PR because review of technical changes can be done independently on translation.

Co-authored-by: Jan Černý <jc@ucw.cz>
Co-authored-by: Vincent Lequertier <vincent@fsfe.org>
Co-authored-by: Greenscreener <git@grsc.cz>
Reviewed-on: https://git.fsfe.org/pmpc/website/pulls/432
Co-authored-by: jc <jc@fsfe.org>
Co-committed-by: jc <jc@fsfe.org>